fix compil. warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Analysis/trunk@174625 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 7ad0befd5f24c2cba2bd70ccb33f09588cd1ea7c

+++ Analysis/Phys/DecayTreeTuple/src/TupleToolBremInfo.cpp
@@ -93,7 +93,7 @@ StatusCode TupleToolBremInfo::fill(const Particle*,
   filltuple &= tuple->column( prefix+"_BremMultiplicity",  P->info(LHCb::Particle::HasBremAdded,0.)); // replace the multiplicity() method that need fullDST
   //brem info from rerunning the BremAdder, looking at the photons
   //This is not correct when running on an mDST - you look at a subset of the total photons in the event.
-  if((int)  P->info(LHCb::Particle::HasBremAdded,0.) != brem.multiplicity() )
+  if((int)  P->info(LHCb::Particle::HasBremAdded,0.) != (int)brem.multiplicity() )
     Warning("TupleToolBremInfo requires fullDST -  BremP and BremOrigin might not be reliable (Multiplicity is OK)",StatusCode::SUCCESS,5);
   filltuple &= tuple->column( prefix+"_BremP"           , brem.momentum());  // need fullDST 
   filltuple &= tuple->column( prefix+"_BremOrigin"      , brem.referencePoint()); // need fullDST
